Move drawtail methods into Bubble as instance methods
